// apps/ktc-console/src/_core/entities/common/creation-information/types.ts
export interface ICreationInformationData {
  createdAt?: string // 2025-04-15T07:57:13.622Z
  createdBy?: string // TODO: 아이디인지, 이름인지?
  updatedAt?: string // 2025-04-15T07:57:13.622Z
  updatedBy?: string// TODO: 아이디인지, 이름인지?
}

export interface ICreationInformation {
  createdAt: string[]
  createdBy: string
  updatedAt: string[]
  updatedBy: string
}